Basic Ingredient Model
Updated Bootstrap Gem to be Bootstrap Sass. Added Sass stylesheet

Made Ingredient model and migrated it with the database

Added a button on the homepage to get to ingredient list.

Made page that lists all the names of ingredients in the database

Validation is only on the name for right now, but plan to add more